requirements.txt
asgiref==3.8.1
Django==4.2.13
django-filter==24.2
djangorestframework==3.15.2
sqlparse==0.5.0
typing_extensions==4.12.2
django-debug-toolbar==4.2.0
Faker==21.0.0
gunicorn==22.0.0